Fan out submariner controller functions
This moves the different resource generators and functionalities
of the controller/submariner package into separate files for different
logical resources.
