+++ controllers/submariner/np_syncer_resources.go
@@ -15,22 +15,35 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package submariner
 
 import (
 	"strconv"
 
+	"github.com/go-logr/logr"
+	"github.com/submariner-io/submariner/pkg/routeagent_driver/constants"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	submopv1a1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	"github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/controllers/helpers"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/network"
 	"github.com/submariner-io/submariner-operator/pkg/names"
 )
 
-func newNetworkPluginSyncerDeployment(cr *submopv1a1.Submariner, clusterNetwork *network.ClusterNetwork) *appsv1.Deployment {
+func (r *SubmarinerReconciler) reconcileNetworkPluginSyncerDeployment(instance *v1alpha1.Submariner,
+	clusterNetwork *network.ClusterNetwork, reqLogger logr.Logger) (*appsv1.Deployment, error) {
+	// Only OVNKubernetes needs networkplugin-syncer so far
+	if instance.Status.NetworkPlugin == constants.NetworkPluginOVNKubernetes {
+		return helpers.ReconcileDeployment(instance, newNetworkPluginSyncerDeployment(instance,
+			clusterNetwork), reqLogger, r.client, r.scheme)
+	}
+	return nil, nil
+}
+
+func newNetworkPluginSyncerDeployment(cr *v1alpha1.Submariner, clusterNetwork *network.ClusterNetwork) *appsv1.Deployment {
 	labels := map[string]string{
 		"app":       "submariner-networkplugin-syncer",
 		"component": "networkplugin-syncer",
