+++ controllers/submariner/submariner_controller.go
@@ -21,26 +21,19 @@ package submariner
 import (
 	"context"
 	"reflect"
-	"strconv"
 
 	"github.com/go-logr/logr"
-	errorutil "github.com/pkg/errors"
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/util/retry"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
@@ -48,16 +41,12 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	submopv1a1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
-	"github.com/submariner-io/submariner-operator/controllers/helpers"
-	"github.com/submariner-io/submariner-operator/controllers/metrics"
 	submarinerclientset "github.com/submariner-io/submariner-operator/pkg/client/clientset/versioned"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/network"
 	"github.com/submariner-io/submariner-operator/pkg/gateway"
 	"github.com/submariner-io/submariner-operator/pkg/images"
-	"github.com/submariner-io/submariner-operator/pkg/names"
 	crdutils "github.com/submariner-io/submariner-operator/pkg/utils/crds"
 	submv1 "github.com/submariner-io/submariner/pkg/apis/submariner.io/v1"
-	"github.com/submariner-io/submariner/pkg/routeagent_driver/constants"
 )
 
 const (
@@ -205,522 +194,6 @@ func (r *SubmarinerReconciler) Reconcile(ctx context.Context, request reconcile.
 	}
 
 	return reconcile.Result{}, nil
-}
-
-func buildGatewayStatusAndUpdateMetrics(gateways *[]submv1.Gateway) []submv1.GatewayStatus {
-	var gatewayStatuses = []submv1.GatewayStatus{}
-
-	if gateways != nil {
-		recordGateways(len(*gateways))
-		// Clear the connections so we don’t remember stale status information
-		recordNoConnections()
-		for _, gateway := range *gateways {
-			gatewayStatuses = append(gatewayStatuses, gateway.Status)
-			recordGatewayCreationTime(gateway.Status.LocalEndpoint, gateway.CreationTimestamp.Time)
-
-			for j := range gateway.Status.Connections {
-				recordConnection(
-					gateway.Status.LocalEndpoint,
-					gateway.Status.Connections[j].Endpoint,
-					string(gateway.Status.Connections[j].Status),
-				)
-			}
-		}
-	} else {
-		recordGateways(0)
-		recordNoConnections()
-	}
-
-	return gatewayStatuses
-}
-
-func (r *SubmarinerReconciler) updateDaemonSetStatus(ctx context.Context, daemonSet *appsv1.DaemonSet, status *submopv1a1.DaemonSetStatus,
-	namespace string) error {
-	if daemonSet != nil {
-		if status == nil {
-			status = &submopv1a1.DaemonSetStatus{}
-		}
-		status.Status = &daemonSet.Status
-		if status.LastResourceVersion != daemonSet.ObjectMeta.ResourceVersion {
-			// The daemonset has changed, check its containers
-			mismatchedContainerImages, nonReadyContainerStates, err :=
-				r.checkDaemonSetContainers(ctx, daemonSet, namespace)
-			if err != nil {
-				return err
-			}
-			status.MismatchedContainerImages = mismatchedContainerImages
-			status.NonReadyContainerStates = nonReadyContainerStates
-			status.LastResourceVersion = daemonSet.ObjectMeta.ResourceVersion
-		}
-	}
-	return nil
-}
-
-func (r *SubmarinerReconciler) checkDaemonSetContainers(ctx context.Context, daemonSet *appsv1.DaemonSet,
-	namespace string) (bool, *[]corev1.ContainerState, error) {
-	containerStatuses, err := r.retrieveDaemonSetContainerStatuses(ctx, daemonSet, namespace)
-	if err != nil {
-		return false, nil, err
-	}
-	var containerImageManifest *string = nil
-	var mismatchedContainerImages = false
-	var nonReadyContainerStates = []corev1.ContainerState{}
-	for i := range *containerStatuses {
-		if containerImageManifest == nil {
-			containerImageManifest = &((*containerStatuses)[i].ImageID)
-		} else if *containerImageManifest != (*containerStatuses)[i].ImageID {
-			// Container mismatch
-			mismatchedContainerImages = true
-		}
-		if !*(*containerStatuses)[i].Started {
-			// Not (yet) ready
-			nonReadyContainerStates = append(nonReadyContainerStates, (*containerStatuses)[i].State)
-		}
-	}
-	return mismatchedContainerImages, &nonReadyContainerStates, nil
-}
-
-func (r *SubmarinerReconciler) retrieveDaemonSetContainerStatuses(ctx context.Context, daemonSet *appsv1.DaemonSet,
-	namespace string) (*[]corev1.ContainerStatus, error) {
-	pods := &corev1.PodList{}
-	selector, err := metav1.LabelSelectorAsSelector(daemonSet.Spec.Selector)
-	if err != nil {
-		return nil, err
-	}
-	err = r.client.List(ctx, pods, client.InNamespace(namespace), client.MatchingLabelsSelector{Selector: selector})
-	if err != nil {
-		return nil, err
-	}
-	containerStatuses := []corev1.ContainerStatus{}
-	for i := range pods.Items {
-		containerStatuses = append(containerStatuses, pods.Items[i].Status.ContainerStatuses...)
-	}
-	return &containerStatuses, nil
-}
-
-func (r *SubmarinerReconciler) retrieveGateways(ctx context.Context, owner metav1.Object, namespace string) (*[]submv1.Gateway, error) {
-	foundGateways := &submv1.GatewayList{}
-	err := r.client.List(ctx, foundGateways, client.InNamespace(namespace))
-	if err != nil && errors.IsNotFound(err) {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	// Ensure we’ll get updates
-	for i := range foundGateways.Items {
-		if err := controllerutil.SetControllerReference(owner, &foundGateways.Items[i], r.scheme); err != nil {
-			return nil, err
-		}
-	}
-	return &foundGateways.Items, nil
-}
-
-func (r *SubmarinerReconciler) reconcileGatewayDaemonSet(
-	instance *submopv1a1.Submariner, reqLogger logr.Logger) (*appsv1.DaemonSet, error) {
-	daemonSet, err := helpers.ReconcileDaemonSet(instance, newGatewayDaemonSet(instance), reqLogger, r.client, r.scheme)
-	if err != nil {
-		return nil, err
-	}
-	err = metrics.Setup(instance.Namespace, instance, daemonSet.GetLabels(), gatewayMetricsServerPort, r.client, r.config, r.scheme, reqLogger)
-	return daemonSet, err
-}
-
-func (r *SubmarinerReconciler) reconcileNetworkPluginSyncerDeployment(instance *submopv1a1.Submariner,
-	clusterNetwork *network.ClusterNetwork, reqLogger logr.Logger) (*appsv1.Deployment, error) {
-	// Only OVNKubernetes needs networkplugin-syncer so far
-	if instance.Status.NetworkPlugin == constants.NetworkPluginOVNKubernetes {
-		return helpers.ReconcileDeployment(instance, newNetworkPluginSyncerDeployment(instance,
-			clusterNetwork), reqLogger, r.client, r.scheme)
-	}
-	return nil, nil
-}
-
-func (r *SubmarinerReconciler) reconcileRouteagentDaemonSet(instance *submopv1a1.Submariner, reqLogger logr.Logger) (*appsv1.DaemonSet,
-	error) {
-	return helpers.ReconcileDaemonSet(instance, newRouteAgentDaemonSet(instance), reqLogger, r.client, r.scheme)
-}
-
-func (r *SubmarinerReconciler) reconcileGlobalnetDaemonSet(instance *submopv1a1.Submariner, reqLogger logr.Logger) (*appsv1.DaemonSet,
-	error) {
-	daemonSet, err := helpers.ReconcileDaemonSet(instance, newGlobalnetDaemonSet(instance), reqLogger, r.client, r.scheme)
-	if err != nil {
-		return nil, err
-	}
-	err = metrics.Setup(instance.Namespace, instance, daemonSet.GetLabels(), globalnetMetricsServerPort,
-		r.client, r.config, r.scheme, reqLogger)
-	return daemonSet, err
-}
-
-func (r *SubmarinerReconciler) serviceDiscoveryReconciler(ctx context.Context, submariner *submopv1a1.Submariner, reqLogger logr.Logger,
-	isEnabled bool) error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if isEnabled {
-			sd := newServiceDiscoveryCR(submariner.Namespace)
-			result, err := controllerutil.CreateOrUpdate(ctx, r.client, sd, func() error {
-				sd.Spec = submopv1a1.ServiceDiscoverySpec{
-					Version:                  submariner.Spec.Version,
-					Repository:               submariner.Spec.Repository,
-					BrokerK8sCA:              submariner.Spec.BrokerK8sCA,
-					BrokerK8sRemoteNamespace: submariner.Spec.BrokerK8sRemoteNamespace,
-					BrokerK8sApiServerToken:  submariner.Spec.BrokerK8sApiServerToken,
-					BrokerK8sApiServer:       submariner.Spec.BrokerK8sApiServer,
-					Debug:                    submariner.Spec.Debug,
-					ClusterID:                submariner.Spec.ClusterID,
-					Namespace:                submariner.Spec.Namespace,
-					GlobalnetEnabled:         submariner.Spec.GlobalCIDR != "",
-					ImageOverrides:           submariner.Spec.ImageOverrides,
-				}
-				if submariner.Spec.CoreDNSCustomConfig != nil {
-					sd.Spec.CoreDNSCustomConfig.ConfigMapName = submariner.Spec.CoreDNSCustomConfig.ConfigMapName
-					sd.Spec.CoreDNSCustomConfig.Namespace = submariner.Spec.CoreDNSCustomConfig.Namespace
-				}
-				if len(submariner.Spec.CustomDomains) > 0 {
-					sd.Spec.CustomDomains = submariner.Spec.CustomDomains
-				}
-				// Set the owner and controller
-				return controllerutil.SetControllerReference(submariner, sd, r.scheme)
-			})
-			if err != nil {
-				return err
-			}
-			if result == controllerutil.OperationResultCreated {
-				reqLogger.Info("Created Service Discovery CR", "Namespace", sd.Namespace, "Name", sd.Name)
-			} else if result == controllerutil.OperationResultUpdated {
-				reqLogger.Info("Updated Service Discovery CR", "Namespace", sd.Namespace, "Name", sd.Name)
-			}
-			return err
-		} else {
-			sdExisting := newServiceDiscoveryCR(submariner.Namespace)
-			err := r.client.Delete(ctx, sdExisting)
-			if errors.IsNotFound(err) {
-				return nil
-			} else if err == nil {
-				reqLogger.Info("Deleted Service Discovery CR", "Namespace", submariner.Namespace)
-			}
-			return err
-		}
-	})
-
-	return errorutil.WithMessagef(err, "error reconciling the Service Discovery CR")
-}
-
-func newGatewayDaemonSet(cr *submopv1a1.Submariner) *appsv1.DaemonSet {
-	labels := map[string]string{
-		"app":       "submariner-gateway",
-		"component": "gateway",
-	}
-
-	revisionHistoryLimit := int32(5)
-
-	maxUnavailable := intstr.FromInt(1)
-
-	deployment := &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels:    labels,
-			Namespace: cr.Namespace,
-			Name:      "submariner-gateway",
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{MatchLabels: map[string]string{"app": "submariner-gateway"}},
-			Template: newGatewayPodTemplate(cr),
-			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
-				RollingUpdate: &appsv1.RollingUpdateDaemonSet{
-					MaxUnavailable: &maxUnavailable,
-				},
-				Type: appsv1.RollingUpdateDaemonSetStrategyType,
-			},
-			RevisionHistoryLimit: &revisionHistoryLimit,
-		},
-	}
-
-	return deployment
-}
-
-// newGatewayPodTemplate returns a submariner pod with the same fields as the cr
-func newGatewayPodTemplate(cr *submopv1a1.Submariner) corev1.PodTemplateSpec {
-	labels := map[string]string{
-		"app": "submariner-gateway",
-	}
-
-	// Create privileged security context for Gateway pod
-	// FIXME: Seems like these have to be a var, so can pass pointer to bool var to SecurityContext. Cleaner option?
-	allowPrivilegeEscalation := true
-	privileged := true
-	runAsNonRoot := false
-	readOnlyRootFilesystem := false
-
-	securityContextAllCapsPrivileged := corev1.SecurityContext{
-		Capabilities:             &corev1.Capabilities{Add: []corev1.Capability{"ALL"}},
-		AllowPrivilegeEscalation: &allowPrivilegeEscalation,
-		Privileged:               &privileged,
-		ReadOnlyRootFilesystem:   &readOnlyRootFilesystem,
-		RunAsNonRoot:             &runAsNonRoot}
-
-	// Create Pod
-	terminationGracePeriodSeconds := int64(1)
-
-	// Default healthCheck Values
-	healthCheckEnabled := true
-	// The values are in seconds
-	healthCheckInterval := uint64(1)
-	healthCheckMaxPacketLossCount := uint64(5)
-
-	if cr.Spec.ConnectionHealthCheck != nil {
-		healthCheckEnabled = cr.Spec.ConnectionHealthCheck.Enabled
-		healthCheckInterval = cr.Spec.ConnectionHealthCheck.IntervalSeconds
-		healthCheckMaxPacketLossCount = cr.Spec.ConnectionHealthCheck.MaxPacketLossCount
-	}
-
-	podTemplate := corev1.PodTemplateSpec{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels: labels,
-		},
-		Spec: corev1.PodSpec{
-			Affinity: &corev1.Affinity{
-				PodAntiAffinity: &corev1.PodAntiAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: []corev1.PodAffinityTerm{{
-						LabelSelector: &metav1.LabelSelector{
-							MatchLabels: labels,
-						},
-						TopologyKey: "kubernetes.io/hostname",
-					}},
-				},
-			},
-			NodeSelector: map[string]string{"submariner.io/gateway": "true"},
-			Containers: []corev1.Container{
-				{
-					Name:            "submariner-gateway",
-					Image:           getImagePath(cr, names.GatewayImage, names.GatewayComponent),
-					ImagePullPolicy: helpers.GetPullPolicy(cr.Spec.Version, cr.Spec.ImageOverrides[names.GatewayComponent]),
-					Command:         []string{"submariner.sh"},
-					SecurityContext: &securityContextAllCapsPrivileged,
-					Env: []corev1.EnvVar{
-						{Name: "SUBMARINER_NAMESPACE", Value: cr.Spec.Namespace},
-						{Name: "SUBMARINER_CLUSTERCIDR", Value: cr.Status.ClusterCIDR},
-						{Name: "SUBMARINER_SERVICECIDR", Value: cr.Status.ServiceCIDR},
-						{Name: "SUBMARINER_GLOBALCIDR", Value: cr.Spec.GlobalCIDR},
-						{Name: "SUBMARINER_CLUSTERID", Value: cr.Spec.ClusterID},
-						{Name: "SUBMARINER_COLORCODES", Value: cr.Spec.ColorCodes},
-						{Name: "SUBMARINER_DEBUG", Value: strconv.FormatBool(cr.Spec.Debug)},
-						{Name: "SUBMARINER_NATENABLED", Value: strconv.FormatBool(cr.Spec.NatEnabled)},
-						{Name: "SUBMARINER_BROKER", Value: cr.Spec.Broker},
-						{Name: "SUBMARINER_CABLEDRIVER", Value: cr.Spec.CableDriver},
-						{Name: "BROKER_K8S_APISERVER", Value: cr.Spec.BrokerK8sApiServer},
-						{Name: "BROKER_K8S_APISERVERTOKEN", Value: cr.Spec.BrokerK8sApiServerToken},
-						{Name: "BROKER_K8S_REMOTENAMESPACE", Value: cr.Spec.BrokerK8sRemoteNamespace},
-						{Name: "BROKER_K8S_CA", Value: cr.Spec.BrokerK8sCA},
-						{Name: "CE_IPSEC_PSK", Value: cr.Spec.CeIPSecPSK},
-						{Name: "CE_IPSEC_DEBUG", Value: strconv.FormatBool(cr.Spec.CeIPSecDebug)},
-						{Name: "SUBMARINER_HEALTHCHECKENABLED", Value: strconv.FormatBool(healthCheckEnabled)},
-						{Name: "SUBMARINER_HEALTHCHECKINTERVAL", Value: strconv.FormatUint(healthCheckInterval, 10)},
-						{Name: "SUBMARINER_HEALTHCHECKMAXPACKETLOSSCOUNT", Value: strconv.FormatUint(healthCheckMaxPacketLossCount, 10)},
-						{Name: "NODE_NAME", ValueFrom: &corev1.EnvVarSource{
-							FieldRef: &corev1.ObjectFieldSelector{
-								FieldPath: "spec.nodeName",
-							},
-						}},
-					},
-				},
-			},
-			// TODO: Use SA submariner-gateway or submariner?
-			ServiceAccountName:            "submariner-gateway",
-			HostNetwork:                   true,
-			TerminationGracePeriodSeconds: &terminationGracePeriodSeconds,
-			RestartPolicy:                 corev1.RestartPolicyAlways,
-			DNSPolicy:                     corev1.DNSClusterFirst,
-			// The gateway engine must be able to run on any flagged node, regardless of existing taints
-			Tolerations: []corev1.Toleration{{Operator: corev1.TolerationOpExists}},
-		},
-	}
-	if cr.Spec.CeIPSecIKEPort != 0 {
-		podTemplate.Spec.Containers[0].Env = append(podTemplate.Spec.Containers[0].Env,
-			corev1.EnvVar{Name: "CE_IPSEC_IKEPORT", Value: strconv.Itoa(cr.Spec.CeIPSecIKEPort)})
-	}
-
-	if cr.Spec.CeIPSecNATTPort != 0 {
-		podTemplate.Spec.Containers[0].Env = append(podTemplate.Spec.Containers[0].Env,
-			corev1.EnvVar{Name: "CE_IPSEC_NATTPORT", Value: strconv.Itoa(cr.Spec.CeIPSecNATTPort)})
-	}
-
-	podTemplate.Spec.Containers[0].Env = append(podTemplate.Spec.Containers[0].Env,
-		corev1.EnvVar{Name: "CE_IPSEC_PREFERREDSERVER", Value: strconv.FormatBool(cr.Spec.CeIPSecPreferredServer)})
-
-	podTemplate.Spec.Containers[0].Env = append(podTemplate.Spec.Containers[0].Env,
-		corev1.EnvVar{Name: "CE_IPSEC_FORCEENCAPS", Value: strconv.FormatBool(cr.Spec.CeIPSecForceUDPEncaps)})
-
-	return podTemplate
-}
-
-func newRouteAgentDaemonSet(cr *submopv1a1.Submariner) *appsv1.DaemonSet {
-	labels := map[string]string{
-		"app":       "submariner-routeagent",
-		"component": "routeagent",
-	}
-
-	matchLabels := map[string]string{
-		"app": "submariner-routeagent",
-	}
-
-	allowPrivilegeEscalation := true
-	privileged := true
-	readOnlyFileSystem := false
-	runAsNonRoot := false
-	securityContextAllCapAllowEscal := corev1.SecurityContext{
-		Capabilities:             &corev1.Capabilities{Add: []corev1.Capability{"ALL"}},
-		AllowPrivilegeEscalation: &allowPrivilegeEscalation,
-		Privileged:               &privileged,
-		ReadOnlyRootFilesystem:   &readOnlyFileSystem,
-		RunAsNonRoot:             &runAsNonRoot,
-	}
-
-	terminationGracePeriodSeconds := int64(1)
-	maxUnavailable := intstr.FromString("100%")
-
-	routeAgentDaemonSet := &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: cr.Namespace,
-			Name:      "submariner-routeagent",
-			Labels:    labels,
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{MatchLabels: matchLabels},
-			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
-				RollingUpdate: &appsv1.RollingUpdateDaemonSet{
-					MaxUnavailable: &maxUnavailable,
-				},
-				Type: appsv1.RollingUpdateDaemonSetStrategyType,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					TerminationGracePeriodSeconds: &terminationGracePeriodSeconds,
-					Volumes: []corev1.Volume{
-						{Name: "host-run", VolumeSource: corev1.VolumeSource{HostPath: &corev1.HostPathVolumeSource{
-							Path: "/run",
-						}}},
-						{Name: "host-sys", VolumeSource: corev1.VolumeSource{HostPath: &corev1.HostPathVolumeSource{
-							Path: "/sys",
-						}}},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:            "submariner-routeagent",
-							Image:           getImagePath(cr, names.RouteAgentImage, names.RouteAgentComponent),
-							ImagePullPolicy: helpers.GetPullPolicy(cr.Spec.Version, cr.Spec.ImageOverrides[names.RouteAgentComponent]),
-							// FIXME: Should be entrypoint script, find/use correct file for routeagent
-							Command:         []string{"submariner-route-agent.sh"},
-							SecurityContext: &securityContextAllCapAllowEscal,
-							VolumeMounts: []corev1.VolumeMount{
-								{Name: "host-sys", MountPath: "/sys", ReadOnly: true},
-								{Name: "host-run", MountPath: "/run"},
-							},
-							Env: []corev1.EnvVar{
-								{Name: "SUBMARINER_NAMESPACE", Value: cr.Spec.Namespace},
-								{Name: "SUBMARINER_CLUSTERID", Value: cr.Spec.ClusterID},
-								{Name: "SUBMARINER_DEBUG", Value: strconv.FormatBool(cr.Spec.Debug)},
-								{Name: "SUBMARINER_CLUSTERCIDR", Value: cr.Status.ClusterCIDR},
-								{Name: "SUBMARINER_SERVICECIDR", Value: cr.Status.ServiceCIDR},
-								{Name: "SUBMARINER_GLOBALCIDR", Value: cr.Spec.GlobalCIDR},
-								{Name: "SUBMARINER_NETWORKPLUGIN", Value: cr.Status.NetworkPlugin},
-								{Name: "NODE_NAME", ValueFrom: &corev1.EnvVarSource{
-									FieldRef: &corev1.ObjectFieldSelector{
-										FieldPath: "spec.nodeName",
-									},
-								}},
-							},
-						},
-					},
-					ServiceAccountName: "submariner-routeagent",
-					HostNetwork:        true,
-					// The route agent engine on all nodes, regardless of existing taints
-					Tolerations: []corev1.Toleration{{Operator: corev1.TolerationOpExists}},
-				},
-			},
-		},
-	}
-
-	return routeAgentDaemonSet
-}
-
-func newGlobalnetDaemonSet(cr *submopv1a1.Submariner) *appsv1.DaemonSet {
-	labels := map[string]string{
-		"app":       "submariner-globalnet",
-		"component": "globalnet",
-	}
-
-	matchLabels := map[string]string{
-		"app": "submariner-globalnet",
-	}
-
-	allowPrivilegeEscalation := true
-	privileged := true
-	readOnlyFileSystem := false
-	runAsNonRoot := false
-	securityContextAllCapAllowEscal := corev1.SecurityContext{
-		Capabilities:             &corev1.Capabilities{Add: []corev1.Capability{"ALL"}},
-		AllowPrivilegeEscalation: &allowPrivilegeEscalation,
-		Privileged:               &privileged,
-		ReadOnlyRootFilesystem:   &readOnlyFileSystem,
-		RunAsNonRoot:             &runAsNonRoot,
-	}
-
-	terminationGracePeriodSeconds := int64(2)
-
-	globalnetDaemonSet := &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: cr.Namespace,
-			Name:      "submariner-globalnet",
-			Labels:    labels,
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{MatchLabels: matchLabels},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name:            "submariner-globalnet",
-							Image:           getImagePath(cr, names.GlobalnetImage, names.GlobalnetComponent),
-							ImagePullPolicy: helpers.GetPullPolicy(cr.Spec.Version, cr.Spec.ImageOverrides[names.GlobalnetComponent]),
-							SecurityContext: &securityContextAllCapAllowEscal,
-							Env: []corev1.EnvVar{
-								{Name: "SUBMARINER_NAMESPACE", Value: cr.Spec.Namespace},
-								{Name: "SUBMARINER_CLUSTERID", Value: cr.Spec.ClusterID},
-								{Name: "SUBMARINER_EXCLUDENS", Value: "submariner-operator,kube-system,operators,openshift-monitoring,openshift-dns"},
-								{Name: "NODE_NAME", ValueFrom: &corev1.EnvVarSource{
-									FieldRef: &corev1.ObjectFieldSelector{
-										FieldPath: "spec.nodeName",
-									},
-								}},
-							},
-						},
-					},
-					ServiceAccountName:            "submariner-globalnet",
-					TerminationGracePeriodSeconds: &terminationGracePeriodSeconds,
-					NodeSelector:                  map[string]string{"submariner.io/gateway": "true"},
-					HostNetwork:                   true,
-					// The Globalnet Pod must be able to run on any flagged node, regardless of existing taints
-					Tolerations: []corev1.Toleration{{Operator: corev1.TolerationOpExists}},
-				},
-			},
-		},
-	}
-
-	return globalnetDaemonSet
-}
-
-func newServiceDiscoveryCR(namespace string) *submopv1a1.ServiceDiscovery {
-	return &submopv1a1.ServiceDiscovery{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: namespace,
-			Name:      names.ServiceDiscoveryCrName,
-		},
-	}
 }
 
 func getImagePath(submariner *submopv1a1.Submariner, imageName, componentName string) string {
